Changes look good in the staging build.

+++ app/src/board/draggable-box.jsx
@@ -152,7 +152,9 @@ class DraggableBox extends React.Component {
   */
   mouseUp = (e) => {
     e.preventDefault();
-    this.props.deleteCallback(this.props.uuid);
+    if (this.props.overDelete) {
+      this.props.deleteCallback(this.props.uuid);
+    };
     this.setState({ draggable: false, resizing: false });
     this.props.callback(this.props.uuid, { curDragging: '' });
     if (!this.state.mouseMoved) {
@@ -190,7 +192,6 @@ class DraggableBox extends React.Component {
         style={this.getBoxStyle()}
       >
         {this.props.children}
-        <div className="pull-tab" style={{ zIndex: this.props.z + 2, opacity: this.state.tabVisible }} />
 
       </div>
     );
